\name{rm.plot-package}
\alias{rm.plot-package}
\docType{package}
\title{
rm.plot: Raw-Means Plots for Experimental Designs
}
\description{
rm.plot (acronym for raw-means plot) is a function for visualizing results of experimental designs with up to two factors (i.e., conditions). The main feature is that it plots both the raw data in the background and superimposed the means to provide a better and more accurate visualization of the underlying distribution. To distinguish between factor levels which would be at the same x-axis position in classical plots, each factor level is offseted from the other factor levels. 
}
\details{
\tabular{ll}{
Package: \tab rm.plot\cr
Type: \tab Package\cr
Version: \tab 0.9.0\cr
Date: \tab 2012-02-20\cr
License: \tab GPL (>= 2)\cr
}
see \code{\link{rm.plot}} for the function, and \code{\link{rm.plot2}} for the wrapper when the data has more than two factors.

see \url{http://www.psychologie.uni-freiburg.de/Members/singmann/R/rm.plot} for theoretical background
}
\author{
Henrik Singmann <henrik.singmann@psychologie.uni-freiburg.de>
}

\keyword{ package }
\seealso{

}

