\name{rm.plot}
\alias{rm.plot}
\title{
rm.plot: Raw-Means Plots for Experimental Designs
}
\description{
rm.plot (for raw-means-plot) is a function for visualizing results of experimental designs with up to two factors. It plots both raw data (background) and factor means (foreground) to provide a more accurate visualization of the underlying distribution.
}
\usage{
rm.plot(d, col.offset = 2, col.x = 3, col.d = 4, noise = 0.008, na.rm = FALSE, 
	pch = 21:25, lty = 1:5, 
	bg.b.col = "darkgrey", bg.f.col = NULL, fg.b.col = "black", fg.f.col = "black", 
	type = "o", pt.cex = 1, xlab = "", ylab = "", 
	ylim, max.offset = 0.2, xaxis = TRUE, x.labels, xaxt = "n", 
	plot = TRUE, legend = TRUE, 
	mar = NULL, reset.mar = TRUE, l.pos, yjust = 0.5, l.bty = "n", l.adj = c(0, 0.5), 
	...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{d}{
	a \code{data.frame} in wide format (see \code{link{reshape}} or the \pkg{reshape} package) that contains at least three columns: one column coding the first factor (\code{col.offset}), one column coding the second factor (\code{col.x}), and one column containing the data (\code{col.d}). The only mandatory argument.
}
  \item{col.offset}{
a \code{numeric} or \code{character} scalar, specifiying either number or name of the column coding the different lines (the offset or first factor). Default is 2.
}
  \item{col.x}{
a \code{numeric} or \code{character} scalar, specifiying either number or name of the column coding the x-axis factor. Default is 3.
}
  \item{col.d}{
a \code{numeric} or \code{character} scalar, specifiying either number or name of the data column. Default is 4.
}
  \item{noise}{
either a \code{numeric} scalar or \code{NULL}. If \code{NULL}, raw data points with same value will be indistinguishable.  Otherwise, uniform noise will be added to each raw data point with same value after the first one with this value. The noise (i.e., jitter) is drawn from the intervall ranging from \code{-noise} to \code{noise} (times the maximal \code{ylim} distance). Default  is .005, that is 0.5% of \code{ylim}.
}
  \item{na.rm}{
\code{logical} indicating whether \code{NA} values should be stripped before the computation proceeds. Default is \code{FALSE}. Throws an error message if FALSE and NAs are encountered.
}
  \item{pch}{
\code{pch} values (plot symbols) taken for plotting the data. Note that the same values are taken for raw data and means. see \code{\link{points}} for more details. Recycled if too short (with warning). Default is 21:25, because those are the only values that can be displayed filled and non-filled. All other values should not be used. 
}
  \item{lty}{
\code{lty} values (line types) for connecting the means. See \code{\link{par}} for more details. Recycled if too short (with warning). Default is 1:5.
}
  \item{bg.b.col}{
background border color: border color of raw data points. Default: \code{"darkgrey"}
}
  \item{bg.f.col}{
background filling color: fill color of raw data points. Default: \code{NULL}
}
  \item{fg.b.col}{
foreground border color: border color of mean data points. Default: \code{black}
}
  \item{fg.f.col}{
foreground fill color: fill color for mean data points. Default: \code{black}
}
  \item{type}{
same as type in \code{\link{plot}}. Default: \code{o} ("overplotted")
}
  \item{pt.cex}{
\code{numeric} specifying the \code{cex} value used for plotting points and lines. Default is 1.
}
  \item{xlab}{
x-axis label. Default: \code{""}
}
  \item{ylab}{
y-axis label. Default: \code{""}
}
  \item{ylim}{
the y-axis limits of the plot. If not specified (the default) will be taken from data so that all raw data points are visible and a warning message is displayed specifying the ylim.
}
  \item{max.offset}{
\code{numeric}. maximal offset of factor levels from the offset factor (\code{col.offset}) specifying the different lines. The centre of each factor on the x-axis is at full numbers (starting from 1 to ...). The maximum will only be reached if the number of factor levels (from \code{col.offset}) is even. Default: 0.2.
}
  \item{xaxis}{
\code{logical} value indicating whether or not the x-axis should be generated by \code{rm.plot}. If \code{TRUE}, labels for the x-axis will be taken either from the unique values of \code{col.x} or can be specified with \code{x.labels}.
}
  \item{x.labels}{
\code{character} vector specifiying \code{col.x} levels. Only relevant if \code{xaxis=TRUE}. Then, the values given here will be displayed at the x-axis for each factor level of \code{col.x}.
}
  \item{xaxt}{
A character which specifies whether ot not the x-axis should be plotted by the call to plot function. Interfers with the aforementioned \code{xaxis} argument and the automatic \code{xaxis} function by \code{rm.plot}. Just there for completeness. Default \code{"n"} (and should not be changed).
}
  \item{plot}{
\code{logical}. Should the \code{rm.plot} be drawn or not. If \code{TRUE} (the default) plot will be drawn. If \code{FALSE} only the legend will be drawn (if \code{legend = TRUE}) See details.
}
  \item{legend}{
\code{logical} indicating whether or not \code{rm.plot} should automatically add a legend on the right outside the plot area indicating which line and points refer to which \code{col.offset} factor levels. Default is \code{TRUE}.
}
  \item{mar}{
\code{NULL} or \code{numerical} vector of length 4 indicating the margins of the plot (see \code{\link{par}}). If \code{NULL} (the default) the right margin (i.e., \code{par("mar")[4]}) will be (imperfectly) guessed from the \code{col.offset} factors for placing the legend right to the plot. If length is four this value will be taken. Ignored for \code{plot = FALSE}.
}
  \item{reset.mar}{
\code{logical} indicating if the margins (\code{mar}) shall be resetted after setting internally. Will be ignored if \code{legend = FALSE}. Default is \code{TRUE} and should not be changed (especially with \code{plot = FLASE}).
}
  \item{l.pos}{
\code{numeric} vector of length 2 indicating the position of the legend. If not specified automatically determined. See details.
}
  \item{yjust}{
how the legend is to be justified relative to the legend y location. A value of 0 means top, 0.5 means centered and 1 means bottom justified. Default is 0.5.
}
  \item{l.bty}{
the type of box to be drawn around the legend. The allowed values are \code{"o"} and \code{"n"} (the default).
}
  \item{l.adj}{
\code{numeric} of length 1 or 2; the string adjustment for legend text. Useful for y-adjustment when labels are plotmath expression. see \code{\link{legend}} and \code{\link{plotmath}} for more info.
}
  \item{\dots}{
further arguments which are either passed to plot or legend. The following argumenst are passed to legend, all others are passed to plot:\cr
\code{"fill", "border", "angle", "density", "box.lwd", "box.lty", "box.col", "pt.cex", "pt.lwd", "xjust", "x.intersp", "y.intersp", "text.width", "text.col", "merge", "trace", "plot", "ncol", "horiz", "title", "inset", "title.col", "title.adj"}
}
}
\details{
\code{rm.plot} is basically an advanced wrapper for two other functions: \code{\link{plot}} and (if \code{legend=TRUE}) \code{\link{legend}}. Furthermore, raw data is plotted with a call to \code{\link{points}} and the means with a call to \code{\link{lines}}.

You can use \code{rm.plot} to plot only a legend by setting \code{plot = FALSE} and \code{legend = TRUE}. Then, \code{rm.plot} will draw an invisible plot with \code{xlim = c(0,10)} and \code{ylim = c(0, 10)} and place the legend on this invisible plot. You can specify \code{l.pos} to position the legend, otherwise it will be plotted at \code{c(5,5)} (i.e., in the middle of the plot). Note that \code{xpd = TRUE} in the call to \code{legend} (see \code{link{par}}).
}
\value{
Nothing
}
\references{
see \url{http://www.psychologie.uni-freiburg.de/Members/singmann/R/rm.plot}
}
\author{
Henrik Singmann
}



\examples{

x <- data.frame(id = 1:150, offset = rep(c("Group A", "Group B", "Group C"),each = 50), xaxis = sample(c("A", "B", "C", "D"),150, replace = TRUE), data = c(rnorm(50, 10, 5), rnorm(50, 15,6), rnorm(50, 20, 5)))

rm.plot(x)

rm.plot(x, main = "Example", ylab = "Values", xlab = "Factor", title = "Groups")

rm.plot(x, "offset", "xaxis", "data")

rm.plot(x, "xaxis", "offset", "data")

rm.plot(x, 3, 2, 4)

x2 <- data.frame(id = 1:150, offset = rep(c("Group A", "Group B", "Group C"),each = 50), xaxis = sample(c("A", "B", "C", "D"),150, replace = TRUE), data = c(rnorm(50, 10, 5), rnorm(50, 15,6), rnorm(50, 20, 5)))

layout(matrix(c(1,2,3,3), 2,2,byrow = TRUE), heights = c(7,1))
rm.plot(x, main = "Data x1", ylab = "Values", xlab = "Factor", legend = FALSE, mar = c(4,4,4,1)+0.1)
rm.plot(x2, main = "Data x2", ylab = "Values", xlab = "Factor", legend = FALSE, mar = c(4,4,4,1)+0.1)
rm.plot(x2, plot = FALSE, title = "Groups")


y <- data.frame(id = 1:300, offset = rep(1, 300), axis = sample(LETTERS[1:6],300, replace = TRUE), data = c(rnorm(100, 1), rnorm(100), rnorm(100,1)))

par(mfrow = c(2,2))

rm.plot(y, legend = FALSE)

rm.plot(y, type = "p", legend = FALSE)

rm.plot(y, type = "l", legend = FALSE)

rm.plot(y, 3, 2, x.labels = "one group only")

 


z <- data.frame (id = 1:200, offset = rep(c("C 1", "C 2"), 200), axis = sample(LETTERS[1:4], 200, replace = TRUE), data = sample(1:20, 200, replace = TRUE))

par(mfrow = c(2,2))

rm.plot(z, noise = NULL, main = "no jitter")

rm.plot(z, noise = .001, main = "smaller jitter")

rm.plot(z, main = "standard jitter")

rm.plot(z, noise = .01, main = "bigger jitter")

}
